agregado caso eliminar tipo productor con productor asociado

+++ src/tests/test_tipo_productor.py
@@ -1,4 +1,4 @@
-from test_model import TipoProductor
+from test_model import TipoProductor, Productor
 from werkzeug.security import  generate_password_hash
 import test_database as test_db
 
@@ -23,14 +23,18 @@ def eliminarTipoProductor(id=-1):
 
     if id == -1:
         return None
-
-    logged_user = test_db.session.query(TipoProductor).filter_by(id = id).first()
-    if logged_user != None:
-        user = test_db.session.query(TipoProductor).get(id)
-        test_db.session.delete(user)
-        test_db.session.commit()
-    else:
+    
+    prod = test_db.session.query(Productor).filter_by(tipo = id).all()
+    if len(prod) > 0:
         return None
+    else:
+        logged_user = test_db.session.query(TipoProductor).filter_by(id = id).first()
+        if logged_user != None:
+            user = test_db.session.query(TipoProductor).get(id)
+            test_db.session.delete(user)
+            test_db.session.commit()
+        else:
+            return None
 
 def agregarTipoProductor(descripcion = ""):
